requirements.txt
groq==0.28.0
sentence-transformers==3.4.1
pinecone-client==5.0.1
openai==1.75.0
pypdf2==3.0.1
streamlit==1.45.0
python-dotenv==1.1.0
requests==2.32.3